Re-factor + change the end points for some commands
